requirements.txt
requests
Pillow
pystray
APScheduler
pywin32
